2715 - Cleaned up the Message container and its item types
